workshops/relational-migration/requirements.txt
chalice
boto3
mysql-connector-python
